feature: Staticmethod support for wrapper 'api_call'
Bump urllib3 to address CVE vulnerabilities + api_call adjustments - when Unauthorized exception is raised, there is an attempt to reconnect to NSX. However it might fail when static method uses that wrapperbecause it tries to read first element of args 'args[0].connection' ('IndexError: out of range' is raised). Staticmethods use connection parameter and it should be accessed through kwargs, not args.

Signed-off-by: Zmijewski, Szymon <szymon.zmijewski@intel.com>